refactor percnotetype map into cpp
