+++ src/musx/dom/PercussionNoteType.h
@@ -22,7 +22,8 @@
 #pragma once
 
 #include <string>
-#include <unordered_map>
+
+#include "musx/dom/Fundamentals.h"
 
 namespace musx {
 namespace dom {
@@ -52,671 +53,7 @@ struct PercussionNoteType
     std::string createName(unsigned orderId) const;
 };
 
-/// @brief C++ version of the `PercNoteTypes.txt` CSV file bundled with the Finale app.
-inline const std::unordered_map<int, PercussionNoteType> percussionNoteTypeMap = {
-// Auxiliary
-    { 6,   { 6,6,"Cowbell",5,56,101,1,false } },
-    { 48,  { 48,6,"Cowbell%g Mute",5,-1,101,1,false } },
-    { 49,  { 49,6,"Cowbell%g Roll",5,-1,101,1,false } },
-    { 50,  { 50,6,"Cowbell%g Tip",5,-1,101,1,false } },
-    { 51,  { 51,6,"Large Cowbell%g",7,-1,101,1,false } },
-    { 52,  { 52,48,"Large Cowbell%g Mute",7,-1,101,1,false } },
-    { 53,  { 53,49,"Large Cowbell%g Roll",7,-1,101,1,false } },
-    { 54,  { 54,50,"Large Cowbell%g Tip",7,-1,101,1,false } },
-    { 55,  { 55,6,"Small Cowbell%g",5,-1,101,1,false } },
-    { 56,  { 56,48,"Small Cowbell%g Mute",5,-1,101,1,false } },
-    { 57,  { 57,49,"Small Cowbell%g Roll",5,-1,101,1,false } },
-    { 58,  { 58,50,"Small Cowbell%g Tip",5,-1,101,1,false } },
-    { 59,  { 59,43,"Shaker",0,82,102,3,false } },
-    { 62,  { 62,59,"Shaker%g Roll",0,-1,102,1,false } },
-    { 63,  { 63,59,"Egg Shaker",0,-1,102,2,false } },
-    { 65,  { 65,63,"Egg Shaker%g Multi Shake",0,-1,102,1,false } },
-    { 522, { 522,522,"Flexatone Low",0,-1,1018,1,false } },
-    { 523, { 523,522,"Flexatone Mid",0,-1,1018,1,false } },
-    { 524, { 524,522,"Flexatone Hi",0,-1,1018,1,false } },
-    { 7,   { 7,7,"Tambourine",0,54,103,2,false } },
-    { 66,  { 66,7,"Tambourine%g Shake",0,-1,103,2,false } },
-    { 67,  { 67,7,"Tambourine%g Thumb Roll",0,-1,103,1,false } },
-    { 8,   { 8,8,"Triangle%g Mute",12,80,104,2,false } },
-    { 9,   { 9,9,"Triangle%g Open",12,81,104,2,false } },
-    { 10,  { 10,10,"High Woodblock",8,76,105,1,false } },
-    { 11,  { 11,10,"Low Woodblock",6,77,105,1,false } },
-    { 68,  { 68,10,"High Jam Block",8,-1,105,1,false } },
-    { 69,  { 69,11,"Low Jam Block",6,-1,105,1,false } },
-    { 70,  { 70,7,"Bell Tree",0,84,1,1,false } },
-    { 511, { 511,7,"Mark Tree",0,-1,1,1,false } },
-    { 71,  { 71,41,"Ratchet",0,-1,1,1,false } },
-    { 72,  { 72,43,"Rattle",0,-1,1,1,false } },
-    { 73,  { 73,251,"Slapstick",0,-1,1,1,false } },
-    { 74,  { 74,7,"Sleigh Bells",0,83,1,1,false } },
-    { 75,  { 75,30,"Spoons",0,-1,1,1,false } },
-    { 12,  { 12,12,"Vibra Slap",0,58,1,1,false } },
-    { 76,  { 76,73,"Whip",0,28,1,1,false } },
-    { 460, { 460,7,"Wind Chimes",0,-1,1,1,false } },
-// Bass Drums
-    { 2,   { 2,2,"Kick Drum",3,36,2,1,false } },
-    { 13,  { 13,2,"Bass Drum",2,35,201,6,true } },
-    { 77,  { 77,13,"Bass Drum%g%d LH",2,-1,201,6,true } },
-    { 78,  { 78,3,"Bass Drum%g%d Rim",2,-1,201,6,true } },
-    { 79,  { 79,3,"Bass Drum%g%d Rim LH",2,-1,201,6,true } },
-    { 80,  { 80,13,"Bass Drum%g%d Roll",2,-1,201,6,true } },
-    { 81,  { 81,13,"Bass Drum%g%d Crush",2,-1,201,6,true } },
-    { 82,  { 82,13,"Bass Drum Unison%g%d Hits",6,-1,220,1,false } },
-    { 83,  { 83,77,"Bass Drum Unison%g%d Hits LH",6,-1,220,1,false } },
-    { 84,  { 84,78,"Bass Drum Unison%g%d Rims",6,-1,220,1,false } },
-    { 85,  { 85,79,"Bass Drum Unison%g%d Rims LH",6,-1,220,1,false } },
-    { 86,  { 86,80,"Bass Drum Unison%g%d Rolls",6,-1,220,1,false } },
-    { 87,  { 87,81,"Bass Drum Unison%g%d Crushes",6,-1,220,1,false } },
-// Cymbals
-    { 4,   { 4,4,"Crash Cymbal",12,49,301,1,false } },
-    { 5,   { 5,4,"Crash Cymbal 2 ",13,57,301,1,false } },
-    { 88,  { 88,4,"Suspended Cymbal%g Roll",12,-1,301,1,false } },
-    { 89,  { 89,88,"Suspended Cymbal%g Roll%d (Mute Release)",12,-1,301,1,false } },
-    { 90,  { 90,4,"Suspended Cymbal%g w/ Mallet%d (Crash)",12,-1,301,1,false } },
-    { 91,  { 91,90,"Suspended Cymbal%g w/ Mallet%d (Loud)",12,-1,301,1,false } },
-    { 92,  { 92,90,"Suspended Cymbal%g w/ Mallet%d (Soft)",12,-1,301,1,false } },
-    { 93,  { 93,88,"Suspended Cymbal%g Cresc%d (Loud)",12,-1,301,1,false } },
-    { 94,  { 94,88,"Suspended Cymbal%g Cresc%d (Soft)",12,-1,301,1,false } },
-    { 95,  { 95,88,"Suspended Cymbal%g Cresc%d (Mute)",12,-1,301,1,false } },
-    { 96,  { 96,93,"Suspended Cymbal%g Cresc Long%d (Loud)",12,-1,301,1,false } },
-    { 97,  { 97,94,"Suspended Cymbal%g Cresc Long%d (Soft)",12,-1,301,1,false } },
-    { 98,  { 98,95,"Suspended Cymbal%g Cresc Long%d (Mute)",12,-1,301,1,false } },
-    { 99,  { 99,93,"Suspended Cymbal%g Cresc Short%d (Loud)",12,-1,301,1,false } },
-    { 100, { 100,94,"Suspended Cymbal%g Cresc Short%d (Soft)",12,-1,301,1,false } },
-    { 101, { 101,95,"Suspended Cymbal%g Cresc Short%d (Mute)",12,-1,301,1,false } },
-    { 102, { 102,4,"Suspended Cymbal%g Bell w/ Shoulder",12,-1,301,1,false } },
-    { 103, { 103,4,"Suspended Cymbal%g Coin Scrape Long",12,-1,301,1,false } },
-    { 104, { 104,103,"Suspended Cymbal%g Coin Scrape Short",12,-1,301,1,false } },
-    { 105, { 105,4,"Suspended Cymbal%g Fat Choke w/ Stick",12,-1,301,1,false } },
-    { 106, { 106,105,"Suspended Cymbal%g Fat Choke w/ Mallet",12,-1,301,1,false } },
-    { 107, { 107,105,"Suspended Cymbal%g Short Choke w/ Stick",12,-1,301,1,false } },
-    { 108, { 108,106,"Suspended Cymbal%g Short Choke w/ Mallet",12,-1,301,1,false } },
-    { 18,  { 18,18,"Ride Cymbal",10,51,302,1,false } },
-    { 19,  { 19,18,"Ride Cymbal%g 2",8,59,302,1,false } },
-    { 109, { 109,18,"Ride Cymbal%g Plastic Tip",10,-1,302,1,false } },
-    { 110, { 110,4,"Ride Cymbal%g Crash",10,-1,302,1,false } },
-    { 20,  { 20,18,"Ride%g Bell",10,53,302,1,false } },
-    { 111, { 111,18,"Ride%g Bell w/ Shoulder",10,-1,302,1,false } },
-    { 14,  { 14,14,"Hi-Hat%g Open",11,46,303,1,false } },
-    { 15,  { 15,15,"Hi-Hat%g Closed",11,42,303,1,false } },
-    { 16,  { 16,15,"Hi-Hat%g Foot",11,44,303,1,false } },
-    { 112, { 112,14,"Hi-Hat%g Foot Splash",11,-1,303,1,false } },
-    { 276, { 276,14,"Hi-Hat%g Half-Open",11,-1,303,1,false } },
-    { 113, { 113,14,"Hi-Hat%g Open w/ Shoulder",11,-1,303,1,false } },
-    { 114, { 114,15,"Hi-Hat%g Closed w/ Shoulder",11,-1,303,1,false } },
-    { 115, { 115,14,"Hi-Hat%g Bell Open",11,-1,303,1,false } },
-    { 116, { 116,15,"Hi-Hat%g Bell Closed",11,-1,303,1,false } },
-    { 117, { 117,15,"Hi-Hat%g Stand",11,-1,303,1,false } },
-    { 21,  { 21,4,"Splash Cymbal",12,55,304,1,false } },
-    { 118, { 118,90,"Splash Cymbal%g w/ Mallet",12,-1,304,1,false } },
-    { 119, { 119,105,"Splash Cymbal%g Choke",12,-1,304,1,false } },
-    { 120, { 120,106,"Splash Cymbal%g Choke w/ Mallet",12,-1,304,1,false } },
-    { 121, { 121,18,"Sizzle Cymbal",8,-1,310,1,false } },
-    { 122, { 122,18,"Sizzle Cymbal%g w/ Mallet",8,-1,310,1,false } },
-    { 123, { 123,88,"Sizzle Cymbal%g Sustained Roll Long",8,-1,310,1,false } },
-    { 124, { 124,88,"Sizzle Cymbal%g Sustained Roll Short",8,-1,310,1,false } },
-    { 125, { 125,18,"Sizzle Cymbal%g Fingers",8,-1,310,1,false } },
-    { 17,  { 17,17,"China Cymbal",14,52,305,1,false } },
-    { 126, { 126,17,"China Cymbal%g w/ Mallet",14,-1,305,1,false } },
-    { 127, { 127,93,"China Cymbal%g Cresc",14,-1,305,1,false } },
-    { 128, { 128,95,"China Cymbal%g Cresc%d (Mute)",14,-1,305,1,false } },
-    { 129, { 129,96,"China Cymbal%g Cresc Long",14,-1,305,1,false } },
-    { 130, { 130,98,"China Cymbal%g Cresc Long%d (Mute)",14,-1,305,1,false } },
-    { 131, { 131,99,"China Cymbal%g Cresc Short",14,-1,305,1,false } },
-    { 132, { 132,101,"China Cymbal%g Cresc Short%d (Mute)",14,-1,305,1,false } },
-    { 133, { 133,107,"China Cymbal%g Choke w/ Stick",14,-1,305,1,false } },
-    { 134, { 134,108,"China Cymbal%g Choke w/ Mallet",14,-1,305,1,false } },
-    { 135, { 135,136,"Large Gong",6,-1,306,1,false } },
-    { 136, { 136,17,"Medium Gong",6,-1,306,1,false } },
-    { 137, { 137,136,"Small Gong",6,-1,306,1,false } },
-    { 138, { 138,136,"Gong%g Choke",6,-1,306,1,false } },
-    { 139, { 139,127,"Gong%g Roll",6,-1,306,1,false } },
-    { 140, { 140,128,"Gong%g Roll%d - Mute Release",6,-1,306,1,false } },
-    { 141, { 141,127,"Gong%g Roll%d - Natural Release",6,-1,306,1,false } },
-    { 142, { 142,136,"Gong%g Besar",6,-1,306,1,false } },
-    { 143, { 143,136,"Java Gong",6,-1,306,1,false } },
-    { 144, { 144,136,"Luo-Gong",6,-1,306,1,false } },
-    { 145, { 145,136,"Mongolian Gong",6,-1,306,1,false } },
-    { 146, { 146,136,"Thai Gong",6,-1,306,1,false } },
-    { 147, { 147,136,"Water Gong",6,-1,306,1,false } },
-    { 148, { 148,136,"Wind Gong",6,-1,306,1,false } },
-    { 149, { 149,136,"Tam-Tam",6,-1,307,1,false } },
-    { 150, { 150,18,"Tam-Tam%g w/ Triangle Beater",6,-1,307,1,false } },
-    { 151, { 151,18,"Tam-Tam%g w/ Wood Stick",6,-1,307,1,false } },
-    { 152, { 152,149,"Tam-Tam%g Dampened",6,-1,307,1,false } },
-    { 153, { 153,103,"Tam-Tam%g Scrape Long",6,-1,307,1,false } },
-    { 154, { 154,103,"Tam-Tam%g Scrape Medium",6,-1,307,1,false } },
-    { 155, { 155,104,"Tam-Tam%g Scrape Short",6,-1,307,1,false } },
-    { 156, { 156,129,"Tam-Tam%g Cresc Long%d (Loud)",6,-1,307,1,false } },
-    { 157, { 157,131,"Tam-Tam%g Cresc Short%d (Loud)",6,-1,307,1,false } },
-    { 158, { 158,130,"Tam-Tam%g Cresc Long%d (Soft)",6,-1,307,1,false } },
-    { 159, { 159,132,"Tam-Tam%g Cresc Short%d (Soft)",6,-1,307,1,false } },
-    { 160, { 160,149,"Wuhan Tam-Tam",6,-1,307,1,false } },
-    { 161, { 161,149,"Bell Plate",6,-1,307,1,false } },
-    { 162, { 162,149,"Earth Plate",6,-1,307,1,false } },
-    { 163, { 163,149,"Earth Plate Scrape",6,-1,307,1,false } },
-    { 164, { 164,4,"Crash Cymbals%g Crash",6,-1,308,2,false } },
-    { 165, { 165,88,"Crash Cymbals%g Circular Roll",6,-1,308,1,false } },
-    { 166, { 166,164,"Crash Cymbals%g Click",6,-1,308,1,false } },
-    { 167, { 167,106,"Crash Cymbals%g Choke Fat",6,-1,308,1,false } },
-    { 168, { 168,106,"Crash Cymbals%g Choke Secco",6,-1,308,1,false } },
-    { 169, { 169,106,"Crash Cymbals%g Crunch Choke",6,-1,308,1,false } },
-    { 170, { 170,164,"Crash Cymbals%g Ding",6,-1,308,1,false } },
-    { 171, { 171,164,"Crash Cymbals%g Fast Zing",6,-1,308,1,false } },
-    { 172, { 172,164,"Crash Cymbals%g Flat Crash",6,-1,308,1,false } },
-    { 173, { 173,88,"Crash Cymbals%g Flat Roll",6,-1,308,1,false } },
-    { 174, { 174,164,"Crash Cymbals%g Port Crash",6,-1,308,1,false } },
-    { 175, { 175,164,"Crash Cymbals%g Scratch In",6,-1,308,1,false } },
-    { 176, { 176,164,"Crash Cymbals%g Scratch Out",6,-1,308,1,false } },
-    { 177, { 177,164,"Crash Cymbals%g Sizz/Suck",6,-1,308,1,false } },
-    { 178, { 178,164,"Crash Cymbals%g Slow Zing",6,-1,308,1,false } },
-    { 179, { 179,106,"Crash Cymbals%g Tap Choke",6,-1,308,1,false } },
-    { 180, { 180,164,"Crash Cymbals%g Tap Halfway",6,-1,308,1,false } },
-    { 181, { 181,88,"Crash Cymbals%g Tremolo",6,-1,308,1,false } },
-    { 182, { 182,164,"Crash Cymbals%g Vacuum",6,-1,308,1,false } },
-    { 183, { 183,164,"Crash Cymbals%g Whale Call",6,-1,308,1,false } },
-    { 184, { 184,164,"Cymbal Section%g Crash",6,-1,309,1,false } },
-    { 185, { 185,165,"Cymbal Section%g Circular Roll",6,-1,309,1,false } },
-    { 186, { 186,166,"Cymbal Section%g Click",6,-1,309,1,false } },
-    { 187, { 187,167,"Cymbal Section%g Crash Choke Fat",6,-1,309,1,false } },
-    { 188, { 188,168,"Cymbal Section%g Crash Choke Secco",6,-1,309,1,false } },
-    { 189, { 189,169,"Cymbal Section%g Crunch Choke",6,-1,309,1,false } },
-    { 190, { 190,170,"Cymbal Section%g Ding",6,-1,309,1,false } },
-    { 191, { 191,171,"Cymbal Section%g Fast Zing",6,-1,309,1,false } },
-    { 192, { 192,172,"Cymbal Section%g Flat Crash",6,-1,309,1,false } },
-    { 193, { 193,173,"Cymbal Section%g Flat Roll",6,-1,309,1,false } },
-    { 194, { 194,15,"Cymbal Section%g Hi-Hat Choke",6,-1,309,1,false } },
-    { 195, { 195,174,"Cymbal Section%g Port Crash",6,-1,309,1,false } },
-    { 196, { 196,175,"Cymbal Section%g Scratch In",6,-1,309,1,false } },
-    { 197, { 197,176,"Cymbal Section%g Scratch Out",6,-1,309,1,false } },
-    { 198, { 198,177,"Cymbal Section%g Sizz/Suck",6,-1,309,1,false } },
-    { 199, { 199,121,"Cymbal Section%g Sizzle",6,-1,309,1,false } },
-    { 200, { 200,178,"Cymbal Section%g Slow Zing",6,-1,309,1,false } },
-    { 201, { 201,179,"Cymbal Section%g Tap Choke",6,-1,309,1,false } },
-    { 202, { 202,180,"Cymbal Section%g Tap Halfway",6,-1,309,1,false } },
-    { 203, { 203,181,"Cymbal Section%g Tremolo",6,-1,309,1,false } },
-    { 204, { 204,182,"Cymbal Section%g Vacuum",6,-1,309,1,false } },
-    { 205, { 205,183,"Cymbal Section%g Whale Call",6,-1,309,1,false } },
-    { 206, { 206,18,"Finger Cymbals",6,-1,311,1,false } },
-    { 207, { 207,206,"Finger Cymbals%g Scrape",6,-1,311,1,false } },
-    { 208, { 208,206,"Finger Cymbals%g Muted Clap",6,-1,311,1,false } },
-    { 209, { 209,206,"Finger Cymbals%g Edge Against Bell",6,-1,311,1,false } },
-    { 210, { 210,206,"Finger Cymbals%g Clasped",6,-1,311,1,false } },
-// Gamelan
-    { 211, { 211,137,"Kenong",0,-1,401,15,false } },
-    { 212, { 212,24,"Kendhang Ageng",0,-1,402,1,false } },
-    { 213, { 213,25,"Kendhang Ciblon",0,-1,402,1,false } },
-    { 214, { 214,26,"Kendhang Batangan",0,-1,402,1,false } },
-    { 215, { 215,27,"Kendhang Ketipung",0,-1,402,1,false } },
-    { 456, { 456,30,"Kempyang",0,-1,4,1,false } },
-    { 457, { 457,31,"Ketuk",0,-1,4,1,false } },
-// Industrial
-    { 216, { 216,18,"Anvil",0,-1,5,1,false } },
-    { 217, { 217,217,"Bicycle Bell",0,-1,5,1,false } },
-    { 218, { 218,2,"Brake Drum",0,-1,5,1,false } },
-    { 219, { 219,13,"Cannon",0,-1,5,1,false } },
-    { 220, { 220,4,"Garbage Can Lid",0,-1,5,2,false } },
-    { 221, { 221,221,"Horn",0,-1,5,1,false } },
-    { 222, { 222,59,"Jar of Nails",0,-1,5,1,false } },
-    { 223, { 223,27,"Plastic Bucket",0,-1,5,1,false } },
-    { 224, { 224,46,"Police Whistle%g ><",0,-1,501,1,false } },
-    { 225, { 225,47,"Police Whistle%g Short",0,-1,501,1,false } },
-    { 226, { 226,46,"Police Whistle%g Sustained",0,-1,501,1,false } },
-    { 455, { 455,59,"Sand Block",0,-1,5,1,false } },
-    { 227, { 227,88,"Saw Blade",0,-1,5,1,false } },
-    { 464, { 464,463,"Scratch Pull",0,30,5,1,false } },
-    { 463, { 463,59,"Scratch Push",0,29,5,1,false } },
-    { 228, { 228,221,"Siren",0,-1,5,1,false } },
-    { 229, { 229,96,"Thunder Machine",0,-1,5,1,false } },
-    { 230, { 230,3,"Typewriter",0,-1,5,1,false } },
-    { 231, { 231,88,"Wind Machine Fast",0,-1,1019,1,false } },
-    { 516, { 516,231,"Wind Machine Slow",0,-1,1019,1,false } },
-    { 517, { 517,231,"Wind Machine Variable",0,-1,1019,1,false } },
-// Miscellaneous
-    { 22,  { 22,22,"Hand Clap",0,39,6,2,false } },
-    { 234, { 234,22,"Finger Snaps",0,-1,6,1,false } },
-    { 461, { 461,461,"Metronome Click",0,32,6,1,false } },
-    { 462, { 462,461,"Metronome Bell",0,34,6,1,false } },
-    { 235, { 235,235,"Rhythm Cues",0,-1,6,1,false } },
-// Snare Drums
-    { 1,   { 1,1,"Snare Drum",7,38,701,2,false } },
-    { 236, { 236,1,"Snare Drum%g%d LH",7,-1,701,1,false } },
-    { 277, { 277,1,"Snare%g Roll",6,-1,701,2,false } },
-    { 512, { 512,1,"Snare Drum (Snares off)",7,-1,701,2,false } },
-    { 513, { 513,1,"Snare Drum%g%d LH (Snares off)",7,-1,701,1,false } },
-    { 514, { 514,1,"Snare%g Roll (Snares off)",6,-1,701,2,false } },
-    { 239, { 239,277,"Snare%g Buzz Roll",7,-1,701,1,false } },
-    { 278, { 278,277,"Snare%g Sustained Roll",6,-1,701,2,false } },
-    { 240, { 240,1,"Snare%g Cross Shot",7,-1,701,1,false } },
-    { 3,   { 3,1,"Snare%g Cross Stick",7,37,701,1,false } },
-    { 243, { 243,1,"Snare%g Dry Crush",7,-1,701,1,false } },
-    { 244, { 244,1,"Snare%g Fat Crush",7,-1,701,1,false } },
-    { 245, { 245,1,"Snare%g Flam",7,-1,701,1,false } },
-    { 246, { 246,3,"Snare%g Ghost Stroke",7,-1,701,1,false } },
-    { 247, { 247,259,"Snare%g Crush Roll Long",7,-1,701,1,false } },
-    { 248, { 248,1,"Snare%g Crush Roll Short",7,-1,701,1,false } },
-    { 249, { 249,3,"Snare%g Rim",7,-1,701,1,false } },
-    { 250, { 250,3,"Snare%g Rim%d LH",7,-1,701,1,false } },
-    { 251, { 251,1,"Snare%g Rim Shot",7,-1,701,1,false } },
-    { 252, { 252,1,"Snare%g Rim Shot%d LH",7,-1,701,1,false } },
-    { 253, { 253,1,"Snare%g Rim Shot%d (Both Hands)",7,-1,701,1,false } },
-    { 237, { 237,1,"Snare Drum%g Brush",7,93,701,1,false } },
-    { 238, { 238,1,"Snare Drum%g Brush Stir",7,-1,701,1,false } },
-    { 241, { 241,1,"Snare Drum%g Dread",7,-1,701,1,false } },
-    { 242, { 242,1,"Snare Drum%g Dread%d LH",7,-1,701,1,false } },
-    { 254, { 254,1,"Snare Drum%g Rod",7,91,701,1,false } },
-    { 255, { 255,1,"Snare Drum%g Rod%d LH",7,-1,701,1,false } },
-    { 256, { 256,3,"Stick Click",7,31,701,1,false } },
-    { 23,  { 23,1,"Electric Snare Drum",7,40,701,1,false } },
-    { 257, { 257,1,"Snare Section%g%d Hits",6,-1,702,1,false } },
-    { 258, { 258,1,"Snare Section%g%d Hits LH",6,-1,702,1,false } },
-    { 259, { 259,277,"Snare Section%g Buzz Rolls",6,-1,702,1,false } },
-    { 458, { 458,277,"Snare Section%g Sustained Rolls",6,-1,702,1,false } },
-    { 260, { 260,240,"Snare Section%g Cross Shots",6,-1,702,1,false } },
-    { 261, { 261,3,"Snare Section%g Cross Sticks",6,-1,702,1,false } },
-    { 264, { 264,243,"Snare Section%g Dry Crushes",6,-1,702,1,false } },
-    { 265, { 265,244,"Snare Section%g Fat Crushes",6,-1,702,1,false } },
-    { 266, { 266,247,"Snare Section%g Crush Roll Long",6,-1,702,1,false } },
-    { 267, { 267,248,"Snare Section%g Crush Roll Short",6,-1,702,1,false } },
-    { 268, { 268,249,"Snare Section%g Rim Shots",6,-1,702,1,false } },
-    { 269, { 269,250,"Snare Section%g Rim Shots%d LH",6,-1,702,1,false } },
-    { 270, { 270,268,"Snare Section%g Rim Shots%d (Both Hands)",6,-1,702,1,false } },
-    { 271, { 271,252,"Snare Section%g Rims",6,-1,702,1,false } },
-    { 272, { 272,253,"Snare Section%g Rims%d LH",6,-1,702,1,false } },
-    { 262, { 262,241,"Snare Section%g Dreads",6,-1,702,1,false } },
-    { 263, { 263,242,"Snare Section%g Dreads%d LH",6,-1,702,1,false } },
-    { 273, { 273,254,"Snare Section%g Rods",6,-1,702,1,false } },
-    { 274, { 274,255,"Snare Section%g Rods%d LH",6,-1,702,1,false } },
-    { 275, { 275,256,"Stick Clicks",6,-1,702,1,false } },
-    { 279, { 279,1,"Side Drum",6,-1,704,1,false } },
-    { 280, { 280,236,"Side Drum%g%d LH",6,-1,704,1,false } },
-    { 281, { 281,1,"Side Drum%g Roll",6,-1,704,1,false } },
-    { 282, { 282,1,"Pipe Snare Drum",6,-1,7,1,false } },
-// Toms/Tenor Drums
-    { 24,  { 24,24,"High Tom",10,50,801,1,false } },
-    { 283, { 283,24,"High Tom%g%d LH",10,-1,801,1,false } },
-    { 284, { 284,289,"High Tom%g Buzz Roll",10,-1,801,1,false } },
-    { 285, { 285,24,"High Tom%g Dry Crush",10,-1,801,1,false } },
-    { 286, { 286,24,"High Tom%g Flam",10,-1,801,1,false } },
-    { 287, { 287,24,"High Tom%g Shot/Rim",10,-1,801,1,false } },
-    { 288, { 288,24,"High Tom%g Shot/Rim%d LH",10,-1,801,1,false } },
-    { 289, { 289,24,"High Tom%g Short Roll",10,-1,801,1,false } },
-    { 290, { 290,284,"High Tom%g Sustained Buzz Roll",10,-1,801,1,false } },
-    { 25,  { 25,25,"High-Mid Tom",9,48,806,1,false } },
-    { 291, { 291,25,"High-Mid Tom%g%d LH",9,-1,806,1,false } },
-    { 292, { 292,297,"High-Mid Tom%g Buzz Roll",9,-1,806,1,false } },
-    { 293, { 293,25,"High-Mid Tom%g Dry Crush",9,-1,806,1,false } },
-    { 294, { 294,25,"High-Mid Tom%g Flam",9,-1,806,1,false } },
-    { 295, { 295,25,"High-Mid Tom%g Shot/Rim",9,-1,806,1,false } },
-    { 296, { 296,25,"High-Mid Tom%g Shot/Rim%d LH",9,-1,806,1,false } },
-    { 297, { 297,25,"High-Mid Tom%g Short Roll",9,-1,806,1,false } },
-    { 298, { 298,292,"High-Mid Tom%g Sustained Buzz Roll",9,-1,806,1,false } },
-    { 26,  { 26,26,"Low-Mid Tom",8,47,807,1,false } },
-    { 299, { 299,26,"Low-Mid Tom%g%d LH",8,-1,807,1,false } },
-    { 300, { 300,305,"Low-Mid Tom%g Buzz Roll",8,-1,807,1,false } },
-    { 301, { 301,26,"Low-Mid Tom%g Dry Crush",8,-1,807,1,false } },
-    { 302, { 302,26,"Low-Mid Tom%g Flam",8,-1,807,1,false } },
-    { 303, { 303,26,"Low-Mid Tom%g Shot/Rim",8,-1,807,1,false } },
-    { 304, { 304,26,"Low-Mid Tom%g Shot/Rim%d LH",8,-1,807,1,false } },
-    { 305, { 305,26,"Low-Mid Tom%g Short Roll",8,-1,807,1,false } },
-    { 306, { 306,300,"Low-Mid Tom%g Sustained Buzz Roll",8,-1,807,1,false } },
-    { 27,  { 27,27,"Low Tom",6,45,808,1,false } },
-    { 307, { 307,27,"Low Tom%g%d LH",6,-1,808,1,false } },
-    { 308, { 308,313,"Low Tom%g Buzz Roll",6,-1,808,1,false } },
-    { 309, { 309,27,"Low Tom%g Dry Crush",6,-1,808,1,false } },
-    { 310, { 310,27,"Low Tom%g Flam",6,-1,808,1,false } },
-    { 311, { 311,27,"Low Tom%g Shot/Rim",6,-1,808,1,false } },
-    { 312, { 312,27,"Low Tom%g Shot/Rim%d LH",6,-1,808,1,false } },
-    { 313, { 313,27,"Low Tom%g Short Roll",6,-1,808,1,false } },
-    { 314, { 314,308,"Low Tom%g Sustained Buzz Roll",6,-1,808,1,false } },
-    { 28,  { 28,28,"Floor Tom 1",5,43,8,1,false } },
-    { 29,  { 29,29,"Floor Tom 2",4,41,8,1,false } },
-    { 315, { 315,315,"Tenors%g High Rims",7,-1,802,1,false } },
-    { 316, { 316,316,"Tenors%g Low Rims",5,-1,802,1,false } },
-    { 317, { 317,275,"Tenors%g Stick Click",6,-1,802,1,false } },
-    { 318, { 318,26,"Impact Drum%g High",7,-1,803,1,false } },
-    { 319, { 319,27,"Impact Drum%g Low",5,-1,803,1,false } },
-    { 320, { 320,24,"RotoTom%g High",7,-1,804,1,false } },
-    { 321, { 321,25,"RotoTom%g Mid",6,-1,804,1,false } },
-    { 322, { 322,26,"RotoTom%g Low",5,-1,804,1,false } },
-    { 323, { 323,25,"Spock Drum%g%d",6,-1,805,2,true } },
-    { 324, { 324,25,"Spock Drum%g%d LH",6,-1,805,2,true } },
-    { 325, { 325,330,"Spock Drum%g Buzz Roll",6,-1,805,2,true } },
-    { 326, { 326,25,"Spock Drum%g Dry Crush",6,-1,805,2,true } },
-    { 327, { 327,25,"Spock Drum%g Flam",6,-1,805,2,true } },
-    { 328, { 328,25,"Spock Drum%g Rim Shot",6,-1,805,2,true } },
-    { 329, { 329,25,"Spock Drum%g Rim Shot%d LH",6,-1,805,2,true } },
-    { 330, { 330,289,"Spock Drum%g Short Roll",6,-1,805,2,true } },
-    { 331, { 331,325,"Spock Drum%g Sustained Buzz Roll",6,-1,805,2,true } },
-// World 1
-    { 30,  { 30,30,"High Agogo",0,67,901,1,false } },
-    { 332, { 332,30,"High Agogo%g Mute",0,-1,901,1,false } },
-    { 31,  { 31,31,"Low Agogo",0,68,901,1,false } },
-    { 333, { 333,31,"Low Agogo%g Mute",0,-1,901,1,false } },
-    { 506, { 506,333,"Agogo Press Sound",0,-1,901,1,false } },
-    { 504, { 504,30,"High Small Agogo",0,-1,901,1,false } },
-    { 505, { 505,31,"Low Small Agogo",0,-1,901,1,false } },
-    { 507, { 507,332,"Small Agogo Press Sound",0,-1,901,1,false } },
-    { 334, { 334,24,"High Bata",0,-1,902,1,false } },
-    { 335, { 335,287,"High Bata%g Slap",0,-1,902,1,false } },
-    { 336, { 336,24,"High Bata%g Dead Stroke",0,-1,902,1,false } },
-    { 337, { 337,25,"Medium Bata",0,-1,902,1,false } },
-    { 338, { 338,295,"Medium Bata%g Slap",0,-1,902,1,false } },
-    { 339, { 339,25,"Medium Bata%g Dead Stroke",0,-1,902,1,false } },
-    { 340, { 340,27,"Low Bata",0,-1,902,1,false } },
-    { 341, { 341,311,"Low Bata%g Slap",0,-1,902,1,false } },
-    { 342, { 342,27,"Low Bata%g Dead Stroke",0,-1,902,1,false } },
-    { 343, { 343,37,"Berimbau",0,-1,903,1,false } },
-    { 344, { 344,36,"Berimbau High",0,-1,903,1,false } },
-    { 345, { 345,12,"Berimbau Buzz",0,-1,903,1,false } },
-    { 346, { 346,30,"High Bongo Bell",0,-1,904,1,false } },
-    { 347, { 347,30,"High Bongo Bell Mute",0,-1,904,1,false } },
-    { 348, { 348,31,"Low Bongo Bell",0,-1,904,1,false } },
-    { 349, { 349,31,"Low Bongo Bell Mute",0,-1,904,1,false } },
-    { 32,  { 32,32,"High Bongo",0,60,905,1,false } },
-    { 350, { 350,32,"High Bongo%g Mute",0,-1,905,1,false } },
-    { 351, { 351,32,"High Bongo%g Slap",0,-1,905,1,false } },
-    { 503, { 503,32,"High Bongo High Tone",0,-1,905,1,false } },
-    { 33,  { 33,33,"Low Bongo",0,61,905,1,false } },
-    { 352, { 352,33,"Low%g Bongo Mute",0,-1,905,1,false } },
-    { 353, { 353,33,"Low%g Bongo Slap",0,-1,905,1,false } },
-    { 354, { 354,354,"Bowl",0,-1,906,5,false } },
-    { 34,  { 34,34,"Cabasa",0,69,907,1,false } },
-    { 355, { 355,34,"Cabasa%g Long",0,-1,907,1,false } },
-    { 356, { 356,34,"Cabasa%g Multi Shake",0,-1,907,1,false } },
-    { 357, { 357,34,"Cabasa%g Shake",0,-1,907,1,false } },
-    { 358, { 358,34,"Cabasa%g Snap",0,-1,907,1,false } },
-    { 359, { 359,2,"Cajone",0,-1,908,3,false } },
-    { 500, { 500,359,"Cajone%g Corner",0,-1,908,3,false } },
-    { 501, { 501,359,"Cajone%g Dead Stroke",0,-1,908,3,false } },
-    { 502, { 502,359,"Cajone%g Fingertips",0,-1,908,3,false } },
-    { 360, { 360,359,"Cajone%g Low",0,-1,908,3,false } },
-    { 361, { 361,359,"Cajone%g Slap",0,-1,908,3,false } },
-    { 362, { 362,362,"Castanets",0,85,909,1,false } },
-    { 363, { 363,362,"Castanets%g LH",0,-1,909,1,false } },
-    { 364, { 364,362,"Castanets%g Flam",0,-1,909,1,false } },
-    { 365, { 365,362,"Castanets%g Roll",0,-1,909,1,false } },
-    { 366, { 366,6,"Cha Cha Bells",0,-1,910,1,false } },
-    { 367, { 367,48,"Cha Cha Bells%g Mute",0,-1,910,1,false } },
-    { 470, { 470,366,"Cha Cha Bells%g Tip",0,-1,910,1,false } },
-    { 35,  { 35,35,"Claves",0,75,911,2,false } },
-    { 36,  { 36,36,"Conga",8,63,912,1,false } },
-    { 368, { 368,36,"Conga%g Bass Tone",8,-1,912,1,false } },
-    { 37,  { 37,36,"Conga%g Dead Stroke",8,62,912,1,false } },
-    { 369, { 369,36,"Conga%g Slap",8,-1,912,1,false } },
-    { 370, { 370,36,"Requinto",11,-1,912,1,false } },
-    { 371, { 371,370,"Requinto%g Bass Tone",11,-1,912,1,false } },
-    { 372, { 372,370,"Requinto%g Dead Stroke",11,-1,912,1,false } },
-    { 373, { 373,370,"Requinto%g Slap",11,-1,912,1,false } },
-    { 374, { 374,36,"Quinto",10,-1,912,1,false } },
-    { 375, { 375,374,"Quinto%g Bass Tone",10,-1,912,1,false } },
-    { 376, { 376,374,"Quinto%g Dead Stroke",10,-1,912,1,false } },
-    { 377, { 377,374,"Quinto%g Slap",10,-1,912,1,false } },
-    { 38,  { 38,38,"Tumba",6,64,912,1,false } },
-    { 378, { 378,38,"Tumba%g Bass Tone",6,-1,912,1,false } },
-    { 379, { 379,38,"Tumba%g Dead Stroke",6,-1,912,1,false } },
-    { 380, { 380,38,"Tumba%g Slap",6,-1,912,1,false } },
-    { 381, { 381,38,"Super Tumba",4,-1,912,1,false } },
-    { 382, { 382,382,"Super Tumba%g Bass Tone",4,-1,912,1,false } },
-    { 383, { 383,382,"Super Tumba%g Dead Stroke",4,-1,912,1,false } },
-    { 384, { 384,382,"Super Tumba%g Slap",4,-1,912,1,false } },
-    { 39,  { 39,39,"Cuica%g High",0,78,913,1,false } },
-    { 385, { 385,39,"Cuica%g Mid",0,-1,913,1,false } },
-    { 40,  { 40,39,"Cuica%g Low",0,79,913,1,false } },
-    { 386, { 386,27,"Darbuka",0,-1,914,1,false } },
-    { 387, { 387,389,"Darbuka%g Dead Stroke",0,-1,914,1,false } },
-    { 388, { 388,311,"Darbuka%g Slap",0,-1,914,1,false } },
-    { 389, { 389,27,"Djembe",0,-1,915,2,false } },
-    { 498, { 498,389,"Djembe%g Bass Tone",0,-1,915,2,false } },
-    { 390, { 390,389,"Djembe%g Dead Stroke",0,-1,915,2,false } },
-    { 499, { 499,390,"Djembe%g Fingertips",0,-1,915,2,false } },
-    { 391, { 391,311,"Djembe%g Slap",0,-1,915,2,false } },
-    { 392, { 392,24,"Frame Drum",0,-1,916,1,false } },
-    { 393, { 393,24,"Frame Drum%g High",0,-1,916,1,false } },
-    { 394, { 394,25,"Frame Drum%g Low",0,-1,916,1,false } },
-    { 395, { 395,26,"Frame Drum%g Bass Tone",0,-1,916,1,false } },
-    { 396, { 396,287,"Frame Drum%g Slap",0,-1,916,1,false } },
-    { 397, { 397,34,"Gourd",0,-1,917,2,false } },
-    { 41,  { 41,41,"Guiro%g Long",0,74,918,1,false } },
-    { 42,  { 42,41,"Guiro%g Short",0,73,918,1,false } },
-    { 471, { 471,41,"Guiro%g Up Long",0,-1,918,1,false } },
-    { 472, { 472,41,"Guiro%g Up Short",0,-1,918,1,false } },
-    { 473, { 473,41,"Guiro%g Down Up",0,-1,918,1,false } },
-    { 398, { 398,41,"Metal Guiro%g Long",0,-1,918,1,false } },
-    { 399, { 399,41,"Metal Guiro%g Short",0,-1,918,1,false } },
-    { 400, { 400,41,"Plastic Guiro%g Long",0,-1,918,1,false } },
-    { 401, { 401,41,"Plastic Guiro%g Short",0,-1,918,1,false } },
-    { 402, { 402,36,"Log Drum%g High",0,-1,919,1,false } },
-    { 403, { 403,38,"Log Drum%g Low",0,-1,919,1,false } },
-// World 2
-    { 43,  { 43,43,"Maracas",0,70,1001,2,false } },
-    { 404, { 404,43,"Hi Maraca",0,-1,1001,1,false } },
-    { 405, { 405,43,"Low Maraca",0,-1,1001,1,false } },
-    { 481, { 481,7,"Pandeiro%g Open",0,-1,1016,2,false } },
-    { 482, { 482,7,"Pandeiro%g Open Fingertips",0,-1,1016,2,false } },
-    { 483, { 483,7,"Pandeiro%g Open Slap",0,-1,1016,2,false } },
-    { 484, { 484,7,"Pandeiro%g Open Thumb",0,-1,1016,2,false } },
-    { 485, { 485,4103,"Pandeiro%g Closed",0,-1,1016,2,false } },
-    { 486, { 486,4103,"Pandeiro%g Closed Fingertips",0,-1,1016,2,false } },
-    { 487, { 487,4103,"Pandeiro%g Closed Slap",0,-1,1016,2,false } },
-    { 488, { 488,4103,"Pandeiro%g Closed Thumb",0,-1,1016,2,false } },
-    { 489, { 489,66,"Pandeiro%g Jingles",0,-1,1016,2,false } },
-    { 406, { 406,392,"Pandero",0,-1,1002,1,false } },
-    { 407, { 407,394,"Pandero%g Dead Stroke",0,-1,1002,1,false } },
-    { 408, { 408,396,"Pandero%g Slap",0,-1,1002,1,false } },
-    { 409, { 409,43,"Rainstick",0,-1,1003,2,false } },
-    { 476, { 476,44,"Repinique",0,-1,1015,1,false } },
-    { 477, { 477,468,"Repinique%g Fingers",0,-1,1015,1,false } },
-    { 478, { 478,467,"Repinique%g Hand",0,-1,1015,1,false } },
-    { 479, { 479,442,"Repinique%g Rim",0,-1,1015,1,false } },
-    { 480, { 480,444,"Repinique%g Slap",0,-1,1015,1,false } },
-    { 410, { 410,34,"Shekere%g High",0,-1,1004,1,false } },
-    { 411, { 411,34,"Shekere%g High Shake",0,-1,1004,1,false } },
-    { 412, { 412,34,"Shekere%g Low",0,-1,1004,1,false } },
-    { 413, { 413,34,"Shekere%g Low Shake",0,-1,1004,1,false } },
-    { 414, { 414,34,"Shekere%g Short",0,-1,1004,1,false } },
-    { 508, { 508,34,"Shekere%g Bass Tone",0,-1,1004,1,false } },
-    { 509, { 509,34,"Shekere%g Down",0,-1,1004,1,false } },
-    { 510, { 510,34,"Shekere%g Up",0,-1,1004,1,false } },
-    { 415, { 415,2,"Surdu",0,87,1005,3,false } },
-    { 416, { 416,81,"Surdu%g Dead Stroke",0,86,1005,3,false } },
-    { 474, { 474,81,"Surdu%g Dead Stroke (Hand)",0,-1,1005,3,false } },
-    { 475, { 475,78,"Surdu%g Rim",0,-1,1005,3,false } },
-    { 417, { 417,36,"Tabla",0,-1,1006,1,false } },
-    { 418, { 418,37,"Tabla%g High",0,-1,1006,1,false } },
-    { 419, { 419,36,"Tabla%g Low",0,-1,1006,1,false } },
-    { 420, { 420,38,"Tabla%g Bass Tone",0,-1,1006,1,false } },
-    { 421, { 421,369,"Tabla%g Slap",0,-1,1006,1,false } },
-    { 422, { 422,24,"Chu",6,-1,1007,2,false } },
-    { 423, { 423,287,"Chu%g Ka (Rim)",6,-1,1007,2,false } },
-    { 424, { 424,25,"Odaiko",6,-1,1007,2,false } },
-    { 425, { 425,295,"Odaiko%g Ka (Rim)",6,-1,1007,2,false } },
-    { 426, { 426,26,"Shime",6,-1,1007,2,false } },
-    { 427, { 427,303,"Shime%g Ka (Rim)",6,-1,1007,2,false } },
-    { 428, { 428,27,"Okedo",6,-1,1007,2,false } },
-    { 429, { 429,311,"Okedo%g Ka (Rim)",6,-1,1007,2,false } },
-    { 430, { 430,28,"Sumo",6,-1,1007,2,false } },
-    { 431, { 431,311,"Sumo%g Ka (Rim)",6,-1,1007,2,false } },
-    { 432, { 432,39,"Talking Drum%g High",0,-1,1008,1,false } },
-    { 433, { 433,40,"Talking Drum%g Low",0,-1,1008,1,false } },
-    { 490, { 490,44,"Tambourim",0,-1,1017,2,false } },
-    { 491, { 491,468,"Tambourim%g Dead Stroke",0,-1,1017,2,false } },
-    { 492, { 492,467,"Tambourim%g Fingertip",0,-1,1017,2,false } },
-    { 493, { 493,442,"Tambourim%g Rim Shot",0,-1,1017,2,false } },
-    { 494, { 494,44,"Tambourim%g Virado Down",0,-1,1017,2,false } },
-    { 495, { 495,468,"Tambourim%g Virado Down Dead Stroke",0,-1,1017,2,false } },
-    { 496, { 496,44,"Tambourim%g Virado Up",0,-1,1017,2,false } },
-    { 497, { 497,468,"Tambourim%g Virado Up Dead Stroke",0,-1,1017,2,false } },
-    { 434, { 434,7,"Tamburin",0,-1,1009,1,false } },
-    { 435, { 435,7,"Tamburin%g High",0,-1,1009,1,false } },
-    { 436, { 436,7,"Tamburin%g Low",0,-1,1009,1,false } },
-    { 437, { 437,7,"Tamburin%g Bass Tone",0,-1,1009,1,false } },
-    { 438, { 438,7,"Tamburin%g Slap",0,-1,1009,1,false } },
-    { 439, { 439,7,"Tamburin%g Jingles",0,-1,1009,1,false } },
-    { 459, { 459,11,"Temple Block Low",0,-1,1014,5,false } },
-    { 518, { 518,459,"Temple Block Low-Mid",0,-1,1014,1,false } },
-    { 519, { 519,459,"Temple Block Mid",0,-1,1014,1,false } },
-    { 520, { 520,459,"Temple Block High-Mid",0,-1,1014,1,false } },
-    { 521, { 521,459,"Temple Block High",0,-1,1014,1,false } },
-    { 440, { 440,36,"Thavil%g Left",0,-1,1010,1,false } },
-    { 441, { 441,38,"Thavil%g Right",0,-1,1010,1,false } },
-    { 44,  { 44,44,"High Timbale",0,65,1011,1,false } },
-    { 444, { 444,44,"High Timbale%g Edge",0,-1,1011,1,false } },
-    { 467, { 467,44,"High Timbale%g Hand",0,-1,1011,1,false } },
-    { 468, { 468,44,"High Timbale%g Mute",0,-1,1011,1,false } },
-    { 442, { 442,44,"High Timbale%g Rim",0,-1,1011,1,false } },
-    { 45,  { 45,44,"Low Timbale",0,66,1011,1,false } },
-    { 465, { 465,44,"Low Timbale%g Edge",0,-1,1011,1,false } },
-    { 466, { 466,45,"Low Timbale%g Hand",0,-1,1011,1,false } },
-    { 469, { 469,45,"Low Timbale%g Mute",0,-1,1011,1,false } },
-    { 443, { 443,44,"Low Timbale%g Rim",0,-1,1011,1,false } },
-    { 445, { 445,30,"Timbale%g Bell",0,-1,1011,1,false } },
-    { 446, { 446,30,"Timbale%g Bell Mute",0,-1,1011,1,false } },
-    { 447, { 447,37,"Udu%g High",0,-1,1012,1,false } },
-    { 448, { 448,36,"Udu%g Low",0,-1,1012,1,false } },
-    { 449, { 449,38,"Udu%g Bass Tone",0,-1,1012,1,false } },
-    { 450, { 450,37,"Udu%g Slap",0,-1,1012,1,false } },
-    { 46,  { 46,46,"Whistle%g Long",0,72,1013,2,false } },
-    { 451, { 451,46,"Whistle%g Medium",0,-1,1013,2,false } },
-    { 47,  { 47,46,"Whistle%g Short",0,71,1013,2,false } },
-    { 452, { 452,7,"Body Bells",0,-1,10,1,false } },
-    { 453, { 453,228,"Bullroarer",0,-1,10,1,false } },
-    { 454, { 454,22,"Clogs",0,-1,10,1,false } },
-// Custom Note Types
-    { 3968, { 3968,3968,"Custom 1",0,0,11,1,false } },
-    { 3969, { 3969,3969,"Custom 2",0,1,11,1,false } },
-    { 3970, { 3970,3970,"Custom 3",0,2,11,1,false } },
-    { 3971, { 3971,3971,"Custom 4",0,3,11,1,false } },
-    { 3972, { 3972,3972,"Custom 5",0,4,11,1,false } },
-    { 3973, { 3973,3973,"Custom 6",0,5,11,1,false } },
-    { 3974, { 3974,3974,"Custom 7",0,6,11,1,false } },
-    { 3975, { 3975,3975,"Custom 8",0,7,11,1,false } },
-    { 3976, { 3976,3976,"Custom 9",0,8,11,1,false } },
-    { 3977, { 3977,3977,"Custom 10",0,9,11,1,false } },
-    { 3978, { 3978,3978,"Custom 11",0,10,11,1,false } },
-    { 3979, { 3979,3979,"Custom 12",0,11,11,1,false } },
-    { 3980, { 3980,3980,"Custom 13",0,12,11,1,false } },
-    { 3981, { 3981,3981,"Custom 14",0,13,11,1,false } },
-    { 3982, { 3982,3982,"Custom 15",0,14,11,1,false } },
-    { 3983, { 3983,3983,"Custom 16",0,15,11,1,false } },
-    { 3984, { 3984,3984,"Custom 17",0,16,11,1,false } },
-    { 3985, { 3985,3985,"Custom 18",0,17,11,1,false } },
-    { 3986, { 3986,3986,"Custom 19",0,18,11,1,false } },
-    { 3987, { 3987,3987,"Custom 20",0,19,11,1,false } },
-    { 3988, { 3988,3988,"Custom 21",0,20,11,1,false } },
-    { 3989, { 3989,3989,"Custom 22",0,21,11,1,false } },
-    { 3990, { 3990,3990,"Custom 23",0,22,11,1,false } },
-    { 3991, { 3991,3991,"Custom 24",0,23,11,1,false } },
-    { 3992, { 3992,3992,"Custom 25",0,24,11,1,false } },
-    { 3993, { 3993,3993,"Custom 26",0,25,11,1,false } },
-    { 3994, { 3994,3994,"Custom 27",0,26,11,1,false } },
-    { 3995, { 3995,3995,"Custom 28",0,27,11,1,false } },
-    { 3996, { 3996,76,"Custom 29",0,-1,11,1,false } },
-    { 3997, { 3997,463,"Custom 30",0,-1,11,1,false } },
-    { 3998, { 3998,464,"Custom 31",0,-1,11,1,false } },
-    { 3999, { 3999,256,"Custom 32",0,-1,11,1,false } },
-    { 4000, { 4000,461,"Custom 33",0,-1,11,1,false } },
-    { 4001, { 4001,4001,"Custom 34",0,33,11,1,false } },
-    { 4002, { 4002,462,"Custom 35",0,-1,11,1,false } },
-    { 4003, { 4003,59,"Custom 36",0,-1,11,1,false } },
-    { 4004, { 4004,74,"Custom 37",0,-1,11,1,false } },
-    { 4005, { 4005,70,"Custom 38",0,-1,11,1,false } },
-    { 4006, { 4006,362,"Custom 39",0,-1,11,1,false } },
-    { 4007, { 4007,416,"Custom 40",0,-1,11,1,false } },
-    { 4008, { 4008,415,"Custom 41",0,-1,11,1,false } },
-    { 4009, { 4009,4009,"Custom 42",0,88,11,1,false } },
-    { 4010, { 4010,4010,"Custom 43",0,89,11,1,false } },
-    { 4011, { 4011,4011,"Custom 44",0,90,11,1,false } },
-    { 4012, { 4012,254,"Custom 45",0,-1,11,1,false } },
-    { 4013, { 4013,4013,"Custom 46",0,92,11,1,false } },
-    { 4014, { 4014,237,"Custom 47",0,-1,11,1,false } },
-    { 4015, { 4015,4015,"Custom 48",0,94,11,1,false } },
-    { 4016, { 4016,4016,"Custom 49",0,95,11,1,false } },
-    { 4017, { 4017,4017,"Custom 50",0,96,11,1,false } },
-    { 4018, { 4018,4018,"Custom 51",0,97,11,1,false } },
-    { 4019, { 4019,4019,"Custom 52",0,98,11,1,false } },
-    { 4020, { 4020,4020,"Custom 53",0,99,11,1,false } },
-    { 4021, { 4021,4021,"Custom 54",0,100,11,1,false } },
-    { 4022, { 4022,4022,"Custom 55",0,101,11,1,false } },
-    { 4023, { 4023,4023,"Custom 56",0,102,11,1,false } },
-    { 4024, { 4024,4024,"Custom 57",0,103,11,1,false } },
-    { 4025, { 4025,4025,"Custom 58",0,104,11,1,false } },
-    { 4026, { 4026,4026,"Custom 59",0,105,11,1,false } },
-    { 4027, { 4027,4027,"Custom 60",0,106,11,1,false } },
-    { 4028, { 4028,4028,"Custom 61",0,107,11,1,false } },
-    { 4029, { 4029,4029,"Custom 62",0,108,11,1,false } },
-    { 4030, { 4030,4030,"Custom 63",0,109,11,1,false } },
-    { 4031, { 4031,4031,"Custom 64",0,110,11,1,false } },
-    { 4032, { 4032,4032,"Custom 65",0,111,11,1,false } },
-    { 4033, { 4033,4033,"Custom 66",0,112,11,1,false } },
-    { 4034, { 4034,4034,"Custom 67",0,113,11,1,false } },
-    { 4035, { 4035,4035,"Custom 68",0,114,11,1,false } },
-    { 4036, { 4036,4036,"Custom 69",0,115,11,1,false } },
-    { 4037, { 4037,4037,"Custom 70",0,116,11,1,false } },
-    { 4038, { 4038,4038,"Custom 71",0,117,11,1,false } },
-    { 4039, { 4039,4039,"Custom 72",0,118,11,1,false } },
-    { 4040, { 4040,4040,"Custom 73",0,119,11,1,false } },
-    { 4041, { 4041,4041,"Custom 74",0,120,11,1,false } },
-    { 4042, { 4042,4042,"Custom 75",0,121,11,1,false } },
-    { 4043, { 4043,4043,"Custom 76",0,122,11,1,false } },
-    { 4044, { 4044,4044,"Custom 77",0,123,11,1,false } },
-    { 4045, { 4045,4045,"Custom 78",0,124,11,1,false } },
-    { 4046, { 4046,4046,"Custom 79",0,125,11,1,false } },
-    { 4047, { 4047,4047,"Custom 80",0,126,11,1,false } },
-    { 4048, { 4048,4048,"Custom 81",0,127,11,1,false } },
-    { 4049, { 4049,4049,"Custom 82",0,-1,11,1,false } },
-    { 4050, { 4050,4050,"Custom 83",0,-1,11,1,false } },
-    { 4051, { 4051,4051,"Custom 84",0,-1,11,1,false } },
-    { 4052, { 4052,4052,"Custom 85",0,-1,11,1,false } },
-    { 4053, { 4053,4053,"Custom 86",0,-1,11,1,false } },
-    { 4054, { 4054,4054,"Custom 87",0,-1,11,1,false } },
-    { 4055, { 4055,4055,"Custom 88",0,-1,11,1,false } },
-    { 4056, { 4056,4056,"Custom 89",0,-1,11,1,false } },
-    { 4057, { 4057,4057,"Custom 90",0,-1,11,1,false } },
-    { 4058, { 4058,4058,"Custom 91",0,-1,11,1,false } },
-    { 4059, { 4059,4059,"Custom 92",0,-1,11,1,false } },
-    { 4060, { 4060,4060,"Custom 93",0,-1,11,1,false } },
-    { 4061, { 4061,4061,"Custom 94",0,-1,11,1,false } },
-    { 4062, { 4062,4062,"Custom 95",0,-1,11,1,false } },
-    { 4063, { 4063,4063,"Custom 96",0,-1,11,1,false } },
-    { 4064, { 4064,4064,"Custom 97",0,-1,11,1,false } },
-    { 4065, { 4065,4065,"Custom 98",0,-1,11,1,false } },
-    { 4066, { 4066,4066,"Custom 99",0,-1,11,1,false } },
-    { 4067, { 4067,4067,"Custom 100",0,-1,11,1,false } },
-    { 4068, { 4068,4068,"Custom 101",0,-1,11,1,false } },
-    { 4069, { 4069,4069,"Custom 102",0,-1,11,1,false } },
-    { 4070, { 4070,4070,"Custom 103",0,-1,11,1,false } },
-    { 4071, { 4071,4071,"Custom 104",0,-1,11,1,false } },
-    { 4072, { 4072,4072,"Custom 105",0,-1,11,1,false } },
-    { 4073, { 4073,4073,"Custom 106",0,-1,11,1,false } },
-    { 4074, { 4074,4074,"Custom 107",0,-1,11,1,false } },
-    { 4075, { 4075,4075,"Custom 108",0,-1,11,1,false } },
-    { 4076, { 4076,4076,"Custom 109",0,-1,11,1,false } },
-    { 4077, { 4077,4077,"Custom 110",0,-1,11,1,false } },
-    { 4078, { 4078,4078,"Custom 111",0,-1,11,1,false } },
-    { 4079, { 4079,4079,"Custom 112",0,-1,11,1,false } },
-    { 4080, { 4080,4080,"Custom 113",0,-1,11,1,false } },
-    { 4081, { 4081,4081,"Custom 114",0,-1,11,1,false } },
-    { 4082, { 4082,4082,"Custom 115",0,-1,11,1,false } },
-    { 4083, { 4083,4083,"Custom 116",0,-1,11,1,false } },
-    { 4084, { 4084,4084,"Custom 117",0,-1,11,1,false } },
-    { 4085, { 4085,4085,"Custom 118",0,-1,11,1,false } },
-    { 4086, { 4086,4086,"Custom 119",0,-1,11,1,false } },
-    { 4087, { 4087,4087,"Custom 120",0,-1,11,1,false } },
-    { 4088, { 4088,4088,"Custom 121",0,-1,11,1,false } },
-    { 4089, { 4089,4089,"Custom 122",0,-1,11,1,false } },
-    { 4090, { 4090,4090,"Custom 123",0,-1,11,1,false } },
-    { 4091, { 4091,4091,"Custom 124",0,-1,11,1,false } },
-    { 4092, { 4092,4092,"Custom 125",0,-1,11,1,false } },
-    { 4093, { 4093,4093,"Custom 126",0,-1,11,1,false } },
-    { 4094, { 4094,4094,"Custom 127",0,-1,11,1,false } },
-    { 4095, { 4095,4095,"Custom 128",0,-1,11,1,false } },
-};
-
-/// @brief Used as a sentinal value if an unknown type is requested.
-inline const PercussionNoteType kUnknownPercussionNoteType{
-    0, 0, "Unknown", 0, -1, 0, 1, false
-};
+const PercussionNoteType& getPercussionNoteTypeFromId(PercussionNoteTypeId percNoteId);
 
 } // namespace percussion
 } // namespace dom
